refactor _ServiceBackend
